requirements.txt
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.4
django-heroku==0.3.1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.0.0
djangorestframework-stubs==0.1.0
firebase-admin==2.14.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
msgpack==0.6.0
Naked==0.1.31
names==0.3.0
protobuf==3.6.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycrypto==2.6.1
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==4.0
shellescape==3.4.1
six==1.12.0
urllib3==1.24.1
virtualenv==16.1.0
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
